Implement MediatR
