+++ Backend/SecretCode.Api/Data/SecretCodeDataContext.cs
@@ -1,0 +1,34 @@
+﻿using Microsoft.EntityFrameworkCore;
+using SecretCode.Api.Models;
+
+namespace SecretCode.Api.Data;
+
+public class SecretCodeDataContext: DbContext
+{
+    #region Constructors
+    public SecretCodeDataContext(DbContextOptions<SecretCodeDataContext> options): base(options)
+    {
+    }
+
+    #endregion
+
+    #region Entities
+    DbSet<User> Users { get; set; }
+    DbSet<Ingredient> Ingredients { get; set; }
+    #endregion
+
+    #region Override Methods
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+
+        modelBuilder.Entity<User>()
+            .Property(_ => _.Name)
+            .IsRequired();
+        
+        modelBuilder.Entity<User>()
+            .HasKey(_ => _.Id);
+    }
+    #endregion    
+}
+
+
