+++ Backend/SecretCode.Api/Models/Ingredient.cs
@@ -1,0 +1,6 @@
+﻿namespace SecretCode.Api.Models;
+
+public class Ingredient : BaseModel
+{
+    public string Name { get; set; }
+}
