+++ Backend/SecretCode.Api/Models/User.cs
@@ -1,0 +1,7 @@
+﻿namespace SecretCode.Api.Models;
+
+public class User: BaseModel
+{
+    public string Name { get; set; }
+    public string Email { get; set; }
+}
