+++ Backend/SecretCode.Api/Repositories/Interfaces/IUserRepository.cs
@@ -1,0 +1,8 @@
+﻿using SecretCode.Api.Models;
+
+namespace SecretCode.Api.Repositories.Interfaces;
+
+public interface IUserRepository: IBaseRepository<User>
+{
+
+}
