+++ Backend/SecretCode.Api/SecretCode.Api.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+  <PropertyGroup>
+    <TargetFramework>net7.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="7.0.12"/>
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0"/>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.13"/>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="7.0.13"/>
+    <PackageReference Include="MediatR" Version="12.1.1"/>
+    <PackageReference Include="AutoMapper" Version="12.0.1"/>
+  </ItemGroup>
+</Project>
