+++ Backend/SecretCode.sln
@@ -1,0 +1,26 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SecretCode.Api", "SecretCode.Api\SecretCode.Api.csproj", "{8F40AF27-0AD0-4C12-A923-FB37D9C50C50}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8F40AF27-0AD0-4C12-A923-FB37D9C50C50}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8F40AF27-0AD0-4C12-A923-FB37D9C50C50}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8F40AF27-0AD0-4C12-A923-FB37D9C50C50}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8F40AF27-0AD0-4C12-A923-FB37D9C50C50}.Release|Any CPU.Build.0 = Release|Any CPU
+		{57F4C36E-43E9-49B4-A18A-F82C75DE9B10}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{57F4C36E-43E9-49B4-A18A-F82C75DE9B10}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{57F4C36E-43E9-49B4-A18A-F82C75DE9B10}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{57F4C36E-43E9-49B4-A18A-F82C75DE9B10}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
