--- SecretCode.Api/Program.cs
@@ -1,35 +1,0 @@
-using SecretCode.Repository;
-using Microsoft.EntityFrameworkCore;
-using SecretCode.Repository.Interfaces;
-
-var builder = WebApplication.CreateBuilder(args);
-ConfigurationManager configuration = builder.Configuration;
-// Add services to the container.
-
-builder.Services.AddControllers();
-// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
-builder.Configuration.AddJsonFile("appsettings.json", optional: true, reloadOnChange: true);
-builder.Configuration.AddJsonFile($"appsettings.{builder.Environment.EnvironmentName}.json", optional: true);
-//Register Services
-builder.Services.AddDbContext<SecretCodeDataContext>(options => 
-    options.UseSqlServer(configuration.GetConnectionString("AwsDatabase")));
-builder.Services.AddScoped<IUserRepository, UserRepository>();
-
-var app = builder.Build();
-
-// Configure the HTTP request pipeline.
-if (app.Environment.IsDevelopment())
-{
-    app.UseSwagger();
-    app.UseSwaggerUI();
-}
-
-app.UseHttpsRedirection();
-
-app.UseAuthorization();
-
-app.MapControllers();
-
-app.Run();
