--- SecretCode.Model/User.cs
@@ -1,7 +1,0 @@
-﻿namespace SecretCode.Model;
-
-public class User: BaseModel
-{
-    public string Name { get; set; }
-    public string Email { get; set; }
-}
