--- SecretCode.Repository/Interfaces/IUserRepository.cs
@@ -1,8 +1,0 @@
-﻿using SecretCode.Model;
-
-namespace SecretCode.Repository.Interfaces;
-
-public interface IUserRepository: IBaseRepository<User>
-{
-
-}
