--- SecretCode.Repository/SecretCode.Repository.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.13"/>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="7.0.13"/>
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\SecretCode.Model\SecretCode.Model.csproj"/>
-  </ItemGroup>
-</Project>
