--- SecretCode.sln
@@ -1,44 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31903.59
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SecretCode.Api", "SecretCode.Api\SecretCode.Api.csproj", "{8F40AF27-0AD0-4C12-A923-FB37D9C50C50}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SecretCode.Model", "SecretCode.Model\SecretCode.Model.csproj", "{4CE37BCC-9E7F-4E19-A549-6C862B963A3F}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SecretCode.Dto", "SecretCode.Dto\SecretCode.Dto.csproj", "{EAB75970-05E9-41C1-B560-031947C69BD3}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SecretCode.Repository", "SecretCode.Repository\SecretCode.Repository.csproj", "{9FFF7DEE-552A-40E1-AAF1-6301AD06178E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8F40AF27-0AD0-4C12-A923-FB37D9C50C50}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8F40AF27-0AD0-4C12-A923-FB37D9C50C50}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8F40AF27-0AD0-4C12-A923-FB37D9C50C50}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8F40AF27-0AD0-4C12-A923-FB37D9C50C50}.Release|Any CPU.Build.0 = Release|Any CPU
-		{57F4C36E-43E9-49B4-A18A-F82C75DE9B10}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{57F4C36E-43E9-49B4-A18A-F82C75DE9B10}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{57F4C36E-43E9-49B4-A18A-F82C75DE9B10}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{57F4C36E-43E9-49B4-A18A-F82C75DE9B10}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4CE37BCC-9E7F-4E19-A549-6C862B963A3F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4CE37BCC-9E7F-4E19-A549-6C862B963A3F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4CE37BCC-9E7F-4E19-A549-6C862B963A3F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4CE37BCC-9E7F-4E19-A549-6C862B963A3F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EAB75970-05E9-41C1-B560-031947C69BD3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EAB75970-05E9-41C1-B560-031947C69BD3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EAB75970-05E9-41C1-B560-031947C69BD3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EAB75970-05E9-41C1-B560-031947C69BD3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9FFF7DEE-552A-40E1-AAF1-6301AD06178E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9FFF7DEE-552A-40E1-AAF1-6301AD06178E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9FFF7DEE-552A-40E1-AAF1-6301AD06178E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9FFF7DEE-552A-40E1-AAF1-6301AD06178E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
